Rename styles.ts to cascade.ts
The module is the cascade: origins, matching, computed values, the
CSSOM behind them. "Styles" read as the stylesheet strings, which live
in useragent.ts.
