+++ tests/presentational-hints.test.ts
@@ -15,7 +15,7 @@
  */
 import {test, expect} from "@b9g/libuild/test";
 import {createDocumentWindow} from "../src/internal/termdom.js";
-import {StyleManager} from "../src/internal/styles.js";
+import {StyleManager} from "../src/internal/cascade.js";
 import {LayoutEngine} from "../src/internal/layout.js";
 import {UA_DOCUMENT_STYLES} from "../src/internal/useragent.js";
